day-5-meta-llama-opensource/requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-readers-file
llama-index-embeddings-ollama